move SpatialMetricsDisplay to utils
